refactor: rename components and optimise the meta tags
