+++ styles/MemberCard.module.css
@@ -42,7 +42,3 @@
   width: 100%;
   height: 380px;
 }
-
-.card__title {
-  font-weight: 500;
-}
